添加测试：Given位数为50, When求值,Then返回12586269025L

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/FibonacciGenerator.java
@@ -1,7 +1,7 @@
 package cc.xpbootcamp.warmup.fibonacci;
 
 public class FibonacciGenerator {
-    public int getFibonacci(int i) {
+    public long getFibonacci(int i) {
         if (i == 1 || i == 2)
             return 1;
         return getFibonacci(i - 1) + getFibonacci(i - 2);
